+++ src/api/projects.rs
@@ -20,6 +20,7 @@ pub mod jobs;
 pub mod labels;
 pub mod members;
 pub mod merge_requests;
+pub mod merge_trains;
 pub mod milestones;
 pub mod pipelines;
 mod project;